server/github: move sync_external_org_with_repo_and_issue to issue service
